auto ownership is simulating
right now only using the first 8 rows of the spec, but we're successfully simulating choices

This works by taking the UEC spreadsheet and keeping the coefficients almost formatted as-is.  I say almost because the filters and alt names had to be tweaked a bit to match Pandas conventions.  This commit is working but brings up a few questions that I will post as issues on github in a bit.  This is clearly not a final implementation but something to elicit early feedback.