adding columns (now 88)  to the auto ownership model
the big change here is that a lot of the columns could not be expressed as DataFrame.eval calls and so I'm directly making Python "eval" calls if the expression starts with an "@".  This is a proposal and not necessarily permanent functionality - something else to talk about.  I like the way this works though as it puts quite a bit of power directly in the spreadsheet for the model specs, much the same way it is now.  We have the option of never doing a DataFrame.eval call but these calls are a bit easier to express and are apparently much faster.

This checkin also has the performance improvements on the choices and converts windows newlines to regular newlines so you can see the spec on github now