got through the first 40 variables or so in the tour freq spec
And all the alternatives, model, and all the new columns that are required to support it.  Only about halfway through at this point but hopefully gets a bit faster going forward.

Also added proper newlines to the csv file so they look ok outside of excel.

Also started keeping a todo list in markdown to keep track of anything that will need revisiting before the model is fully usable.