basic formulation of the mode choice model
Really just pulling in all the right pieces - this is the plumbing bit, though it is technically already making choices, but using exactly one variable.

For now this captures the nesting structure and has stacked the spec so there aren't 18 columns.  This is a big step forward, but now the fun begins.  There are multiple non-totally-solved issues including 1) what to do in the stacked format when an expression/coefficient applies to more than one alternative and 2) what to do about all the coefficients being set to "1" and then moving the coefficients into the expression.