Add a utility that converts utilities to probabilities
Picking and choosing code from UrbanSim's mnl module.
This utils_to_probs function assumes that data and coefficients
have already been combined to give a table of choosers
and their utilities across alternatives. This function then
exponentiates the utilities and normalizes them for each chooser
so that they can be treated as probabilities.