a pretty large commit for mode choice
The biggest change here is that there is now a skims object which manages the numerous skim objects.  I already found at least one way in which this is not sufficient so more changes will be coming.  But this does get all the other models working with the global skims object.

This also allows for expressions in the coefficients - this works by calling "eval" on all of the expressions in order so for instance the coefficient for in vehicle time is set as a float first and then many other coefficients are set with the relative relationship to ivt.  There are still numerous changes to come, but I don't want to keep this much code without checking in.