this checkin adds skim3d object
This is used to select between skim objects on a row-by-row basis in python.

Before we were dereferencing a matrix of impedances with a Series of origins and a Series of destinations.  We now dereference a third dimension which is most frequently used for time periods, so that tours that start at different times can reference (quickly) different skim matrices.

This is implemented internally by having a 3d matrix where the 3rd dimension is the time period (or something else).  Of course all that complexity is hidden from the user.