Add back new simple_simulate function
This function is intended for situations in which there no interaction
terms between alternatives and choosers and in which there is no
sampling of alternatives. The only data necessary in this situation
are choosers and a model specification/utilities, plus optional
skim data. The procedure for handling this situation is so straightforward
that it's useful to treat it separately.