size terms are now fairly complete
Size terms are added as a table in the tables directory with one helper function which has a unit test under defaults.  Size terms are now used in the workplace location and destination choice models.  The destination choice models no longer use sampling, so are fairly close to final models.