large advance on mandatory scheduling
Only a few "windowing" variables are left unimplemented.

This implements all of the mandatory scheduling spec otherwise, as well as vectorizing the trip scheduling (doing the 1st trips, 2nd trips, etc for each person) and also fills in the start_previous and end_prevoius variables for the previous trip for each person.  This functionality is now a separate utility function in the util directory and has doc strings and unit tests now.