reverse the order of the tests so that the mini_run test passes when run by itself
basically there's still some issue with results being cached when running simulations within the same process.  since this really isn't the intended use case of the simulation framework (it only really comes up in tests), I don't think we should worry about it too much.  I'm just going to run the test that looks at specific results as the first test so that the results won't change because of the simulation that comes before it.