organizing all the segments so then can be columns in a stacked dataframe
Turns out the mode choice models for different purposes have basically the same form (with small exceptions for the rescord and at work mode choice).  I wrote a script that stacks the specs and lines them up so that now all the mode choice models can be in the same csv file.  This should be a major preventative step against copy and paste errors as the expressions now apply to all the segments (to have different expressions, have multiple rows with zeros where you don't want the utility to apply).  There are still a number of steps remaining, but this is a good first step.