am now stacking the config anymore
This was going to be a huge pain as I went to the later variables, where an expression would be used for 4-5 different alternatives.  This would result in a lot of copy and paste and the potential for error.  Instead, I'm stacking BUT the alternative column is now a comma separted list of alternative names.  This seems like a good compromise to me.  So now every expression has a single row, but you can also put multiple segments in the same csv file as separate columns.

So that's exactly what I did.  I had to do some serious pandas-fu on the stack_mode_choice reprocessing script, but I think the data format is very easy to interpret.  Anyway, the csv now has separate columns for separate trip purposes and only one row per expression.  Very nice.

I also realized that the actual coefficients vary by segment (duh) and so there is now a second csv file with all of the coefficient expressions for each segment.  This was in the YAML before but I think it makes sense to put it in a csv file since you can see the segments side-by-side (as a table).  Will have to get some feedback on this though.

About 54 rows have been implemented and it's time to start on that again.

Nothing is unit tested at this point so that will have to come as well.  Although the Pandas operations are fairly complicated, the file formats are pretty darn simple, so hopefully it's a good solution for the user.