I think I'm going to leave performance inprovements here
The main strategy here is to keep track of which columns are dependent on which model results - for instance hhsize ?== auto_ownership is dependent on the auto_ownership model and only add those columns to the larger persons dataframe after the model runs.  This gives a performance improvement of about 30% or so at the cost of the model telling the sim framework exactly when to compute which column.