refactor: better checks and warnings when SPEC or COEFFICENT fields are missing values.
because the fields in the settings classes are inherited, it is not
usually possible for the settings checker to determine if a path to an
external file is actually required for a particular model component.

the workaround is to issue a specific warning that a path *may* be
expected and users should check the YAML file.

the ultimate solution would be to define more robust Pydantic data
models that ensure that fields are marked as required when appropriate.